+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Changed
 
 - The `Tlv` and `Lv` API return `&[u8]` instead of `Option<&[u8]>`.
+- `ByteConversionError` error variants `ToSliceTooSmall` and `FromSliceTooSmall` are struct
+  variants now. `SizeMissmatch` was removed appropriately.
 
 ## Added
 
@@ -24,6 +26,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 # Added
 
 - Added `pdu_datafield_len` getter function for `PduHeader`
+
+## Removed
+
+- `SizeMissmatch` because it is not required for the `ByteConversionError` error enumeration
+  anymore.
 
 # [v0.7.0-beta.0] 2023-08-16